extract build_images
