+++ fpr/pipelines/run_repo_tasks.py
@@ -33,6 +33,7 @@ import fpr.docker.volumes as volumes
 from fpr.models.pipeline import Pipeline
 from fpr.models.org_repo import OrgRepo
 from fpr.models.git_ref import GitRef
+from fpr.docker.images import build_images
 from fpr.models.language import (
     ContainerTask,
     DependencyFile,
@@ -352,24 +353,6 @@ def iter_task_envs(
             package_manager.tasks[task_name] for task_name in args.repo_task
         ]
         yield language, package_manager, image, version_commands, tasks
-
-
-async def build_images(
-    docker_pull: bool, images: Iterable[DockerImage]
-) -> Iterable[str]:
-    try:
-        built_image_tags: Iterable[str] = await asyncio.gather(
-            *[
-                containers.build(
-                    image.dockerfile_bytes, image.local.repo_name, pull=docker_pull
-                )
-                for image in images
-            ]
-        )
-        return built_image_tags
-    except Exception as err:
-        log.error(f"error occurred building images: {err}\n{exc_to_str()}")
-        raise err
 
 
 async def run_pipeline(
